Explicitly state where static middleware should be used
I'm not sure if the way I've done it is correct. Initially, my frontend worked only on localhost and not on Heroku. I fixed it by placing static middleware after cors middleware. I think it makes sense.  So I added this to make it clear for others, who may have the same problem.